Rich text: extract input and selection
